Project directory refactor and reorganization
