+++ src/semantics/SemanticAnalyzer.cpp
@@ -1,4 +1,4 @@
-#include "include/SemanticAnalyzer.h"
+#include "semantics/SemanticAnalyzer.h"
 
 #include <string>
 #include <vector>
